% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMSSB.R
\name{AMSSB}
\alias{AMSSB}
\title{Fit an additive model with discrete spike and slab prior in variational Bayes.}
\usage{
AMSSB(
  X,
  Y,
  degree = 5,
  knots = NULL,
  Boundary.knots = NULL,
  spline_der_length = 1000,
  svd_threshold = 0.01,
  ...
)
}
\arguments{
\item{X}{An n by G matrix of covariates.}

\item{Y}{The outcome vector with length n, or the n by 1 outcome matrix.}

\item{degree}{The degree of the spline basis. Default is 5.}

\item{knots}{The knots of the spline basis. Default is NULL, which means that the knots are equally spaced.}

\item{Boundary.knots}{The boundary knots of the spline basis. Default is NULL, which means that the boundary knots are the minimum and maximum of the covariates.}

\item{spline_der_length}{The total length of intervals for the spline second derivative. Default is 1000.}

\item{svd_threshold}{The threshold for the singular value decomposition of the kernel matrix. Any singular value below this threshold will be set to this threshold. Default is 0.01.}

\item{...}{Other arguments used in GVSSB.}
}
\value{
An object with S3 class "AMSSB".
\tabular{ll}{
   \code{data} \tab The data information. \cr
   \tab \cr
   \code{beta} \tab The sparse estimation of coefficients. \cr
   \tab \cr
   \code{intercept} \tab The fitted value of intercept. \cr
   \tab \cr
   \code{sigma_noise} \tab The fitted value of noise level \eqn{\tilde{\sigma}}.\cr
   \tab \cr
   \code{mu} \tab The posterior mean of the coefficients. \cr
   \tab \cr
   \code{Sigma} \tab The posterior covariance matrix of the coefficients. \cr
   \tab \cr
   \code{gamma} \tab The posterior inclusion probability for each groups.\cr
   \tab \cr
   \code{selected_groups} \tab The positive group indexes selected by the algorithm.\cr
   \tab \cr
   \code{lambda} \tab The value of the hyperparameter \eqn{\lambda}. If em=FALSE, this value is the same as the input argument.\cr
   \tab \cr
   \code{prior} \tab The slab part of the coefficient prior.\cr
   \tab \cr
   \code{nu} \tab Degree of freedom when the slab is T distribution.\cr
}
}
\description{
This function gives a sparse estimation of the coefficients in additive model using spike and slab variational Bayes. To make prediction, please use \code{\link{predict.AMSSB}}.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 300
X <- mvtnorm::rmvnorm(n, sigma=diag(G))

# define additive model
f1 <- function(x) 5 * sin(x)
f2 <- function(x) 2 * (x^2 - 0.5)
f3 <- function(x) 2 * exp(x)
f4 <- function(x) 3 * x

# generate response vector
Y <- f1(X[,1]) + f2(X[,2]) + f3(X[,3]) + f4(X[,4]) + rnorm(n, 0, 1)

# fit AMSSB model
model.Laplace <- AMSSB(X, Y, prior = 'Laplace')

}
\seealso{
\code{\link{GVSSB}}, \code{\link{predict.AMSSB}}, \code{\link{cv.GVSSB}}
}
