% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca1chr}
\alias{brca1chr}
\title{Breast cancer gene expression data from The Cancer Genome Atlas with genes grouped by chromosome segments.}
\format{
A list with 3 elements:
\itemize{
  \item X: gene expression data for 4704 genes and 536 samples.
  \item y: brca1 gene expression data for 536 samples.
  \item groups: group index for each gene in X annotated by chromosome segments.
}
}
\usage{
brca1chr
}
\description{
A dataset containing the gene expression data for 4704 genes and the brca1 gene from 536 samples.
}
\keyword{datasets}
