% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Gives the coverage of a fitted GVSSB model.}
\usage{
coverage(object, alpha = 0.95, beta = NULL, type = c("group", "marginal"))
}
\arguments{
\item{object}{A fitted GVSSB object.}

\item{alpha}{A numeric value between 0 and 1 indicating the confidence level. Default is 0.95.}

\item{beta}{A numeric vector of coefficients. Default is NULL.}

\item{type}{A character string indicating the type of credible region volume. Default is "group".}
}
\value{
A list of the region volume and coverage (if beta is provided).
}
\description{
This function takes in a fitted GVSSB model and a confidence level, and gives the credible region volume of the fitted model. If a coefficient beta is provided, the coverage of the fitted model is also given. Both group-level and marginal-level credible regions are supported.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 200
p_i <- 5
p <- G * p_i
X <- mvtnorm::rmvnorm(n, sigma=diag(p))

# generate coefficients
k <- 5
beta <- rep(0,p)
nonzero_group <- sample(1:G, k)
for(index in nonzero_group){
    beta[p_i * (index - 1) + 1:p_i] <- runif(p_i, -1, 1)
}

# define group index
groups <- rep(1:G, each = p_i)

# generate response vector

Y <- X \%*\% beta + rnorm(n, 0, sd = 3)

# fit GVSSB model
model.Laplace <- GVSSB(X, Y, groups, prior = 'Laplace')

# get the coverage
coverage.group = coverage(model.Laplace, alpha = 0.95, beta = beta, type = "group")
coverage.marginal = coverage(model.Laplace, alpha = 0.95, beta = beta, type = "marginal")
}
