% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.GVSSB.R
\name{cv.GVSSB}
\alias{cv.GVSSB}
\title{Choose the prior of GVSSB using cross-validation.}
\usage{
cv.GVSSB(X, Y, groups, nfolds = 5, loss = c("L2", "L1"), ...)
}
\arguments{
\item{X}{An n by p matrix of covariates.}

\item{Y}{The outcome vector with length n, or the n by 1 outcome matrix.}

\item{groups}{The group indicator vector with length p.}

\item{nfolds}{The total number of folds for cross-validation. Default is 5.}

\item{loss}{Loss used as cross-validation error. Valid options are "L1" and "L2". Default is "L2".}

\item{...}{Other arguments used in GVSSB.}
}
\value{
An object with S3 class "GVSSB" selected by cross-validation.
\tabular{ll}{
   \code{beta} \tab The sparse estimation of coefficients. \cr
   \tab \cr
   \code{intercept} \tab The fitted value of intercept. \cr
   \tab \cr
   \code{sigma} \tab The fitted value of noise level \eqn{\tilde{\sigma}}.\cr
   \tab \cr
   \code{gamma} \tab The posterior inclusion probability for each groups.\cr
   \tab \cr
   \code{lambda} \tab The value of the hyperparameter \eqn{\lambda}. If em=FALSE, this value is the same as the input argument.\cr
   \tab \cr
   \code{groups_index} \tab The group index with length p.\cr
   \tab \cr
   \code{prior} \tab The slab part of the coefficient prior.\cr
   \tab \cr
   \code{nu} \tab Degree of freedom when the slab is T distribution.\cr
}
}
\description{
This function does k-fold cross-validation for GVSSB and chooses the prior with the smallest cv error. Gaussian prior, Laplace prior and Cauchy prior are considered in this function.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 200
p_i <- 5
p <- G * p_i
X <- mvtnorm::rmvnorm(n, sigma=diag(p))

# generate coefficients
k <- 10
beta <- rep(0,p)
nonzero_group <- sample(1:G, k)
for(index in nonzero_group){
    beta[p_i * (index - 1) + 1:p_i] <- runif(p_i, -1, 1)
}

# define group index
groups <- rep(1:G, each = p_i)

# generate response vector
Y <- X \%*\% beta + rnorm(n, 0, 1)

# fit GVSSB model
fit <- cv.GVSSB(X, Y, groups, nfolds = 5, loss = 'L2')


}
\seealso{
\code{\link{GVSSB}}, \code{\link{predict.GVSSB}}
}
