% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVSSB.R
\name{GVSSB}
\alias{GVSSB}
\title{This function gives a sparse estimation of the coefficients in group linear model using spike and slab variational Bayes.}
\usage{
GVSSB(
  X,
  Y,
  groups,
  w,
  sigma_tilde,
  prior = "Gaussian",
  nu = 1,
  lambda = 1,
  em = T,
  alpha = 0,
  beta = 0,
  threshold = 0.5,
  tol = 1e-05,
  iter.max = 5000
)
}
\arguments{
\item{X}{An n by p matrix of covariates.}

\item{Y}{The outcome vector with length n, or the n by 1 outcome matrix.}

\item{groups}{The group indicator vector with length p.}

\item{w}{The hyperparameter controlling the sparsity.}

\item{sigma_tilde}{The initial value of noise level \eqn{\tilde{\sigma}^2}.}

\item{prior}{The slab part of the coefficient prior. Default is "Gaussian".}

\item{nu}{Degree of freedom when the slab is T distribution. Default is 1.}

\item{lambda}{The initial value of hyperparameter \eqn{\lambda}. Default is 1.}

\item{em}{Whether update hyperparameters using variational EM algorithm. Default is TRUE.}

\item{alpha}{Shape of the inverse gamma distribution for the noise level \eqn{\tilde{\sigma}^2}. Default is 0.}

\item{beta}{Scale of the inverse gamma distribution for the noise level \eqn{\tilde{\sigma}^2}. Default is 0.}

\item{threshold}{Threshold of the posterior inclusion probability. Only coefficients with posterior inclusion larger than this threshold are considered nonzero (signals).}

\item{tol}{Convergence tolerance. The algorithm iterates untill the difference of binary entropy between two consecutive steps is smaller than this value. Default is 1e-5.}

\item{iter.max}{Maximum iteration number of the algorithm. Default is 5000.}
}
\value{
A list of values containing the result of the algorithm:
\tabular{ll}{
   \code{beta} \tab The sparse estimation of coefficients. \cr
   \tab \cr
   \code{intercept} \tab The fitted value of intercept. \cr
   \tab \cr
   \code{sigma} \tab The fitted value of noise level \eqn{\tilde{\sigma}}.\cr
   \tab \cr
   \code{gamma} \tab The posterior inclusion probability for each groups.\cr
   \tab \cr
   \code{lambda} \tab The value of the hyperparameter \eqn{\lambda}. If em=FALSE, this value is the same as the input argument.\cr
}
}
\description{
This function gives a sparse estimation of the coefficients in group linear model using spike and slab variational Bayes.
}
\examples{

data(Birthwt)
X <- Birthwt$X
Y <- Birthwt$bwt
groups <- Birthwt$group

fit.Gaussian <- GVSSB(X, Y, groups, prior = 'Gaussian')
fit.Laplace <- GVSSB(X, Y, groups, prior = 'Laplace')
fit.Cauchy <- GVSSB(X, Y, groups, prior = 'T', nu = 1)

}
