% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.AMSSB.R
\name{plot.AMSSB}
\alias{plot.AMSSB}
\title{Gives the spline fitting plot of an additive model by AMSSB}
\usage{
\method{plot}{AMSSB}(object, index = NULL, subtract_mean = TRUE, length_out = 500, ...)
}
\arguments{
\item{object}{A fitted AMSSB object.}

\item{index}{The index of the covariate to be plotted.}

\item{subtract_mean}{A logical value indicating whether to subtract the mean from the fitted values. Default is TRUE.}

\item{length_out}{The number of points to be used for the spline fitting.}

\item{...}{Additional arguments to be passed to the plot function.}
}
\value{
A plot of the spline fitting for the specified group.
}
\description{
This function takes in a fitted AMSSB model and the group index to be plotted, and generates a spline fitting plot based on the fitted values.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 300
X <- mvtnorm::rmvnorm(n, sigma=diag(G))

# define additive model
f1 <- function(x) 5 * sin(x)
f2 <- function(x) 2 * (x^2 - 0.5)
f3 <- function(x) 2 * exp(x)
f4 <- function(x) 3 * x

# generate response vector
Y <- f1(X[,1]) + f2(X[,2]) + f3(X[,3]) + f4(X[,4]) + rnorm(n, 0, 1)

# fit AMSSB model
model.Laplace <- AMSSB(X, Y, prior = 'Laplace')

# plot spline fitting for group 1
plot.AMSSB(model.Laplace, index = 1)

}
