% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.AMSSB.R
\name{predict.AMSSB}
\alias{predict.AMSSB}
\title{Gives the prediction of an additive model problem using a fitted AMSSB model.}
\usage{
\method{predict}{AMSSB}(object, X, truncate = TRUE, ...)
}
\arguments{
\item{object}{A fitted AMSSB object.}

\item{X}{An n by G matrix of covariates.}

\item{truncate}{A logical value indicating whether to truncate the covariates. Default is TRUE.}

\item{...}{Other arguments used in AMSSB.}
}
\value{
An object with S3 class "predicted_AMSSB".
\tabular{ll}{
   \code{X} \tab The covariate matrix. \cr
   \tab \cr
   \code{Y_hat} \tab The estimated value of the response vector. \cr
   \tab \cr
   \code{model} \tab The fitted AMSSB model. \cr
}
}
\description{
This function takes in a fitted AMSSB model and a covariate matrix, and gives a prediction of the response vector.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 300
X <- mvtnorm::rmvnorm(n, sigma=diag(G))

# define additive model
f1 <- function(x) 5 * sin(x)
f2 <- function(x) 2 * (x^2 - 0.5)
f3 <- function(x) 2 * exp(x)
f4 <- function(x) 3 * x

# generate response vector
Y <- f1(X[,1]) + f2(X[,2]) + f3(X[,3]) + f4(X[,4]) + rnorm(n, 0, 1)

# fit AMSSB model
model.Laplace <- AMSSB(X, Y, prior = 'Laplace')

# generate new data
n_test <- 50
X_test <- mvtnorm::rmvnorm(n_test, sigma = diag(G))

# get the prediction
predict.Laplace <- predict.AMSSB(model.Laplace, X_test)

}
\seealso{
\code{\link{AMSSB}}, \code{\link{cv.AMSSB}}
}
