% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVSSB.R
\name{GVSSB}
\alias{GVSSB}
\title{Fit a group linear regression model with discrete spike and slab prior in variational Bayes.}
\usage{
GVSSB(
  X,
  Y,
  groups,
  Omega,
  w,
  lambda,
  sigma_noise,
  prior = c("Gaussian", "Laplace", "T"),
  nu = 1,
  em = T,
  alpha = 0,
  beta = 0,
  threshold = 0.5,
  tol = 1e-05,
  iter.max = 5000,
  info = T,
  standardize = T
)
}
\arguments{
\item{X}{An n by p matrix of covariates.}

\item{Y}{The outcome vector with length n, or the n by 1 outcome matrix.}

\item{groups}{The group indicator vector with length p.}

\item{Omega}{The prior precision matrix of the coefficients. Default is identity matrix.}

\item{w}{The hyperparameter controlling the sparsity. If missing, the algorithm will use group Lasso to initialize this value.}

\item{lambda}{The initial value of hyperparameter \eqn{\lambda}. If missing, the algorithm will use group Lasso to initialize this value.}

\item{sigma_noise}{The initial value of noise level \eqn{\tilde{\sigma}^2}.}

\item{prior}{The slab part of the coefficient prior. Default is "Gaussian".}

\item{nu}{Degree of freedom when the slab is T distribution. Only useful when prior='T'. Default is 1.}

\item{em}{Whether update hyperparameters using variational EM algorithm. Default is TRUE.}

\item{alpha}{Shape of the inverse gamma distribution for the noise level \eqn{\tilde{\sigma}^2}. Default is 0.}

\item{beta}{Scale of the inverse gamma distribution for the noise level \eqn{\tilde{\sigma}^2}. Default is 0.}

\item{threshold}{Threshold of the posterior inclusion probability. Only coefficients with posterior inclusion larger than this threshold are considered nonzero (signals).}

\item{tol}{Convergence tolerance. The algorithm iterates untill the difference of binary entropy between two consecutive steps is smaller than this value. Default is 1e-5.}

\item{iter.max}{Maximum iteration number of the algorithm. Default is 5000.}

\item{info}{Whether print the iteration numbers per 100 steps and when the convergence is satisfied. Default is TRUE.}

\item{standardize}{Whether standardize the covariates. Default is TRUE.}
}
\value{
An object with S3 class "GVSSB".
\tabular{ll}{
   \code{data} \tab The data information. \cr
   \tab \cr
   \code{beta} \tab The sparse estimation of coefficients. \cr
   \tab \cr
   \code{intercept} \tab The fitted value of intercept. \cr
   \tab \cr
   \code{sigma_noise} \tab The fitted value of noise level \eqn{\tilde{\sigma}}.\cr
   \tab \cr
   \code{mu} \tab The posterior mean of the coefficients. \cr
   \tab \cr
   \code{Sigma} \tab The posterior covariance matrix of the coefficients. \cr
   \tab \cr
   \code{gamma} \tab The posterior inclusion probability for each groups.\cr
   \tab \cr
   \code{selected_groups} \tab The positive group indexes selected by the algorithm.\cr
   \tab \cr
   \code{lambda} \tab The value of the hyperparameter \eqn{\lambda}. If em=FALSE, this value is the same as the input argument.\cr
   \tab \cr
   \code{prior} \tab The slab part of the coefficient prior.\cr
   \tab \cr
   \code{nu} \tab Degree of freedom when the slab is T distribution.\cr
}
}
\description{
This function gives a sparse estimation of the coefficients in group linear model using spike and slab variational Bayes. To make prediction, please use \code{\link{predict.GVSSB}}.
}
\examples{
# load GVSSB library
library(GVSSB)

# generate covariates
n <- 200
G <- 200
p_i <- 5
p <- G * p_i
X <- mvtnorm::rmvnorm(n, sigma=diag(p))

# generate coefficients
k <- 10
beta <- rep(0,p)
nonzero_group <- sample(1:G, k)
for(index in nonzero_group){
    beta[p_i * (index - 1) + 1:p_i] <- runif(p_i, -1, 1)
}

# define group index
groups <- rep(1:G, each = p_i)

# generate response vector

snr <- 1
Y <- X \%*\% beta + rnorm(n, 0, sd = sqrt(var(X \%*\% beta) / snr))

# fit GVSSB model
model.Gaussian <- GVSSB(X, Y, groups, prior = 'Gaussian')
model.Laplace <- GVSSB(X, Y, groups, prior = 'Laplace')
model.Cauchy <- GVSSB(X, Y, groups, prior = 'T', nu = 1)
}
\seealso{
\code{\link{AMSSB}}, \code{\link{predict.GVSSB}}, \code{\link{cv.GVSSB}}
}
